+++ src/lsp/protocol/position.cr
@@ -3,7 +3,7 @@ module LSP
     include JSON::Serializable
 
     # Line position in a document (zero-based).
-    property line : Int32
+    property line : Int64
 
     # Character offset on a line in a document (zero-based). Assuming that
     # the line is represented as a string, the `character` value represents
@@ -11,7 +11,7 @@ module LSP
     #
     # If the character value is greater than the line length it defaults back
     # to the line length.
-    property character : Int32
+    property character : Int64
 
     def initialize(@line, @character)
     end
